feat: HandlerAdapter를 관리하는 HandlerAdapterRepository 추가
