+++ README.md
@@ -22,5 +22,5 @@
 ### 기능 요구사항
 interface 기반의 컨트롤러와 annotation 기반의 컨트롤러가 공존하는 상태로 정상 동작하도록 구현한다.
 - [X] ControllerScanner 클래스에서 @Controller가 붙은 클래스를 찾을 수 있다.
-- [ ] HandlerMappingRegistry 클래스에서 HandlerMapping을 처리하도록 구현했다.
-- [ ] HandlerAdapterRegistry 클래스에서 HandlerAdapter를 처리하도록 구현했다.
+- [X] HandlerMappingRegistry 클래스에서 HandlerMapping을 처리하도록 구현했다.
+- [X] HandlerAdapterRegistry 클래스에서 HandlerAdapter를 처리하도록 구현했다.
