+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -4,16 +4,11 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.lang.reflect.InvocationTargetException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
 import nextstep.mvc.adapter.HandlerAdapter;
+import nextstep.mvc.adapter.HandlerAdapterRepository;
 import nextstep.mvc.mapping.HandlerMapping;
 import nextstep.mvc.mapping.HandlerMappingRepository;
 import nextstep.mvc.view.ModelAndView;
-import org.reflections.Reflections;
-import org.reflections.scanners.Scanners;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -23,36 +18,21 @@ public class DispatcherServlet extends HttpServlet {
     private static final Logger log = LoggerFactory.getLogger(DispatcherServlet.class);
 
     private final HandlerMappingRepository handlerMappingRepository;
-    private final List<HandlerAdapter> handlerAdapters;
+    private final HandlerAdapterRepository handlerAdapterRepository;
 
     public DispatcherServlet() {
         this.handlerMappingRepository = new HandlerMappingRepository();
-        this.handlerAdapters = new ArrayList<>();
+        this.handlerAdapterRepository = new HandlerAdapterRepository();
     }
 
     @Override
     public void init() {
         handlerMappingRepository.init();
-        final Reflections adapterReflections = new Reflections("nextstep.mvc.adapter", Scanners.SubTypes);
-        final Set<Class<? extends HandlerAdapter>> adapterTypes = adapterReflections.getSubTypesOf(
-                HandlerAdapter.class);
-        for (Class<? extends HandlerAdapter> adapterType : adapterTypes) {
-            addAdapterInstances(adapterType);
-        }
+        handlerAdapterRepository.init();
     }
 
     public void addHandlerMapping(final HandlerMapping handlerMapping) {
         handlerMappingRepository.add(handlerMapping);
-    }
-
-    private void addAdapterInstances(final Class<? extends HandlerAdapter> adapterType) {
-        try {
-            final HandlerAdapter handlerAdapter = adapterType.getDeclaredConstructor().newInstance();
-            handlerAdapters.add(handlerAdapter);
-        } catch (InstantiationException | IllegalAccessException |
-                 InvocationTargetException | NoSuchMethodException e) {
-            throw new IllegalArgumentException(e);
-        }
     }
 
     @Override
@@ -62,19 +42,12 @@ public class DispatcherServlet extends HttpServlet {
 
         try {
             final Object handler = handlerMappingRepository.getHandler(request);
-            final HandlerAdapter handlerAdapter = getAdapter(handler);
+            final HandlerAdapter handlerAdapter = handlerAdapterRepository.getAdapter(handler);
             final ModelAndView modelAndView = handlerAdapter.handle(request, response, handler);
             modelAndView.render(request, response);
         } catch (Throwable e) {
             log.error("Exception : {}", e.getMessage(), e);
             throw new ServletException(e.getMessage());
         }
-    }
-
-    private HandlerAdapter getAdapter(final Object handler) {
-        return handlerAdapters.stream()
-                .filter(handlerAdapter -> handlerAdapter.supports(handler))
-                .findAny()
-                .orElseThrow(IllegalArgumentException::new);
     }
 }
